added filters by action name
