--- src/main/java/eosio/spectrum/websocket/api/Message/Account_ram_deltas.java
@@ -1,4 +1,0 @@
-package eosio.spectrum.websocket.api.Message;
-
-public class Account_ram_deltas {
-}
