--- src/main/java/eosio/spectrum/websocket/api/Message/Failed_dtrx_trace.java
@@ -1,4 +1,0 @@
-package eosio.spectrum.websocket.api.Message;
-
-public class Failed_dtrx_trace {
-}
