--- src/main/java/eosio/spectrum/websocket/api/Message/Partial.java
@@ -1,4 +1,0 @@
-package eosio.spectrum.websocket.api.Message;
-
-public class Partial {
-}
