--- src/main/java/eosio/spectrum/websocket/api/Message/Receipt.java
@@ -1,4 +1,0 @@
-package eosio.spectrum.websocket.api.Message;
-
-public class Receipt {
-}
