+++ src/main/java/eosio/spectrum/websocket/api/Message/chronicle/Account_ram_deltas.java
@@ -1,0 +1,4 @@
+package eosio.spectrum.websocket.api.Message.chronicle;
+
+public class Account_ram_deltas {
+}
