+++ src/main/java/eosio/spectrum/websocket/api/Message/chronicle/Act.java
@@ -1,0 +1,7 @@
+package eosio.spectrum.websocket.api.Message.chronicle;
+
+import java.util.List;
+
+public class Act {
+    private List<Authorization> authorization;
+}
