+++ src/main/java/eosio/spectrum/websocket/api/Message/chronicle/Auth_sequence.java
@@ -1,0 +1,6 @@
+package eosio.spectrum.websocket.api.Message.chronicle;
+
+public class Auth_sequence {
+    private int sequence;
+    private String account;
+}
