+++ src/main/java/eosio/spectrum/websocket/api/Message/chronicle/Authorization.java
@@ -1,0 +1,6 @@
+package eosio.spectrum.websocket.api.Message.chronicle;
+
+public class Authorization {
+    private String actor;
+    private String permission;
+}
