+++ src/main/java/eosio/spectrum/websocket/api/Message/chronicle/Failed_dtrx_trace.java
@@ -1,0 +1,4 @@
+package eosio.spectrum.websocket.api.Message.chronicle;
+
+public class Failed_dtrx_trace {
+}
