+++ src/main/java/eosio/spectrum/websocket/api/Message/chronicle/Partial.java
@@ -1,0 +1,4 @@
+package eosio.spectrum.websocket.api.Message.chronicle;
+
+public class Partial {
+}
