+++ src/main/java/eosio/spectrum/websocket/api/Message/chronicle/Transaction.java
@@ -1,4 +1,6 @@
-package eosio.spectrum.websocket.api.Message;
+package eosio.spectrum.websocket.api.Message.chronicle;
+
+import eosio.spectrum.websocket.api.Message.chronicle.Trace;
 
 public class Transaction {
     private int block_num;
